requirements.txt
conan
PySide2
